+++ collect_app/src/main/java/org/odk/collect/android/fragments/dialogs/CustomDatePickerDialog.java
@@ -187,26 +187,26 @@ public abstract class CustomDatePickerDialog extends DialogFragment {
         gregorianDateText.setText(label);
     }
 
-    protected void setUpDayPicker(LocalDateTime ethiopianDate) {
+    protected void setUpDayPicker(LocalDateTime date) {
         dayPicker.setMinValue(1);
-        dayPicker.setMaxValue(ethiopianDate.dayOfMonth().getMaximumValue());
+        dayPicker.setMaxValue(date.dayOfMonth().getMaximumValue());
         if (datePickerDetails.isSpinnerMode()) {
-            dayPicker.setValue(ethiopianDate.getDayOfMonth());
+            dayPicker.setValue(date.getDayOfMonth());
         }
     }
 
-    protected void setUpMonthPicker(LocalDateTime ethiopianDate, String[] monthsArray) {
+    protected void setUpMonthPicker(LocalDateTime date, String[] monthsArray) {
         monthPicker.setMaxValue(monthsArray.length - 1);
         monthPicker.setDisplayedValues(monthsArray);
         if (!datePickerDetails.isYearMode()) {
-            monthPicker.setValue(ethiopianDate.getMonthOfYear() - 1);
+            monthPicker.setValue(date.getMonthOfYear() - 1);
         }
     }
 
-    protected void setUpYearPicker(LocalDateTime ethiopianDate, int minSupportedYear, int maxSupportedYear) {
+    protected void setUpYearPicker(LocalDateTime date, int minSupportedYear, int maxSupportedYear) {
         yearPicker.setMinValue(minSupportedYear);
         yearPicker.setMaxValue(maxSupportedYear);
-        yearPicker.setValue(ethiopianDate.getYear());
+        yearPicker.setValue(date.getYear());
     }
 
     public int getDay() {
