+++ collect_app/src/main/java/org/odk/collect/android/widgets/EthiopianDateWidget.java
@@ -17,14 +17,10 @@
 package org.odk.collect.android.widgets;
 
 import android.content.Context;
-import android.view.View;
 
 import org.javarosa.form.api.FormEntryPrompt;
 import org.odk.collect.android.activities.FormEntryActivity;
 import org.odk.collect.android.fragments.dialogs.EthiopianDatePickerDialog;
-import org.odk.collect.android.utilities.DateTimeUtils;
-
-import java.util.Date;
 
 import static org.odk.collect.android.fragments.dialogs.CustomDatePickerDialog.DATE_PICKER_DIALOG;
 
@@ -35,23 +31,6 @@ public class EthiopianDateWidget extends AbstractDateWidget {
 
     public EthiopianDateWidget(Context context, FormEntryPrompt prompt) {
         super(context, prompt);
-    }
-
-    @Override
-    protected void createWidget() {
-        super.createWidget();
-        dateButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                showDatePickerDialog();
-            }
-        });
-    }
-
-    @Override
-    protected void setDateLabel() {
-        isNullAnswer = false;
-        dateTextView.setText(DateTimeUtils.getDateTimeLabel((Date) getAnswer().getValue(), datePickerDetails, false, getContext()));
     }
 
     protected void showDatePickerDialog() {
