+++ collect_app/src/main/java/org/odk/collect/android/widgets/IslamicDateWidget.java
@@ -17,14 +17,10 @@
 package org.odk.collect.android.widgets;
 
 import android.content.Context;
-import android.view.View;
 
 import org.javarosa.form.api.FormEntryPrompt;
 import org.odk.collect.android.activities.FormEntryActivity;
 import org.odk.collect.android.fragments.dialogs.IslamicDatePickerDialog;
-import org.odk.collect.android.utilities.DateTimeUtils;
-
-import java.util.Date;
 
 import static org.odk.collect.android.fragments.dialogs.CustomDatePickerDialog.DATE_PICKER_DIALOG;
 
@@ -32,23 +28,6 @@ public class IslamicDateWidget extends AbstractDateWidget {
 
     public IslamicDateWidget(Context context, FormEntryPrompt prompt) {
         super(context, prompt);
-    }
-
-    @Override
-    protected void createWidget() {
-        super.createWidget();
-        dateButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                showDatePickerDialog();
-            }
-        });
-    }
-
-    @Override
-    protected void setDateLabel() {
-        isNullAnswer = false;
-        dateTextView.setText(DateTimeUtils.getDateTimeLabel((Date) getAnswer().getValue(), datePickerDetails, false, getContext()));
     }
 
     protected void showDatePickerDialog() {
